feat: Auto-generated entries when using Conventional Commits
Adds the function for Conventional Commits for automated changelog
generation and versioning.
The conventional commits can be read from a file that must be created by
the command 'git log <range> -v'

Depending on the conventional commits the new function can also do an
auto release from 'Unreleased' depending on the conventional commits
that are made.